admin ProtectedRoute and Outlet
