+++ admin/src/post/post_delete.rs
@@ -12,7 +12,6 @@ pub fn PostDeleteForm(id: String, slug: Signal<String>) -> impl IntoView {
     let value = post_delete.value();
 
     create_effect(move |_| {
-        log!("navigate create_effect run");
         let v = value();
         if let Some(v) = v {
             let post_result = v.map_err(|_| PostError::ServerError).flatten();
